+++ NEWS.md
@@ -12,6 +12,7 @@
 * Add support for custom Keycloak base URL for JWKS endpoint, new environment variable `KC_JWKS_BASE_URL` (MODSIDECAR-148)
 * Review and clean up logs in mgr-apllications (MGRAPPS-84)
 * Allow discovery information to be specified for UI modules (MGRAPPS-92)
+* Add API endpoint: Get Application Discovery By Query (MGRAPPS-98)
 ---
 
 ## Version `v3.0.0` (12.03.2025)